Refactors PostService options for remote handling of keys and values.
